Extract the test subscriber so it could be used in other tests
