+++ Libraries/CO.CDP.MQ.Tests/DispatcherContractTest.cs
@@ -1,4 +1,3 @@
-using System.Collections.Concurrent;
 using FluentAssertions;
 
 namespace CO.CDP.MQ.Tests;
@@ -135,28 +134,4 @@ public abstract class DispatcherContractTest
     protected record TestMessage(int Id, String Name);
 
     private record UnexpectedTestMessage(int Id, String Name);
-
-    protected class TestSubscriber<TEvent>(Func<TEvent, Task> handler)
-        : ISubscriber<TEvent> where TEvent : class
-    {
-        public readonly ConcurrentBag<object> HandledMessages = new();
-        public readonly ConcurrentBag<object> ReceivedMessages = new();
-
-        public TestSubscriber(CancellationTokenSource? cancellationTokenSource = null) : this(
-            _ =>
-            {
-                cancellationTokenSource?.Cancel();
-                return Task.CompletedTask;
-            })
-        {
-        }
-
-        public Task Handle(TEvent @event)
-        {
-            ReceivedMessages.Add(@event);
-            var result = handler(@event);
-            HandledMessages.Add(@event);
-            return result;
-        }
-    }
 }
